state of the app
This is the current development stage of the app. the "enciclopedia" is the main engine for retrieving information and for landing pages of entities.

Orientation and matching are features for contributions of data and matching the data to our geo-entities. This is not so styled nor debugged nor fully developed yet.

Most actual code is AI generated.